Revert "sna/gen4+: Backport tight vertex packing for simple renderblits"
This reverts commit 8ff76fad1fadc5e309f9a12c30f883460a432049 and
commit 48e4dc4bd4b2980f0f804f572d0e3fc1bb4bc21e.

I forgot gen4 and gen5 do not have the 'non-normalized' bit in their
sampler states.

+++ src/sna/sna_render.h
@@ -126,8 +126,8 @@ struct sna_composite_op {
 		} gen4;
 
 		struct {
-			int wm_kernel;
-			int ve_id;
+			int16_t wm_kernel;
+			int16_t ve_id;
 		} gen5;
 
 		struct {
